Prevent unnecessary re-renders for TodosView
Say we are on the list (filter) of archived todos, and we add a todo.
We don't need to render again the list of todos because the added todo
was only for the list (filter) of 'active items'.

One exception, if the current filter is for the 'active todos', in this
case we need to allow a new render of the component because we want to
see all the todos with the new one.