Rewrite todos' state change
Previously all items were browsed while we were looking for the item to
modify, which is totally useless.
Now the item to modify is directly found, updated, and replaced.
This should improve performances in some cases (if lot of items).

Also, some examples of code are left as it, but commented, just for the
demonstration.